Fix bugs with non-English characters
We were being sloppy in various places about using `len()` to get string
lengths instead of `chars().count()`; also we now normalize Unicode
combining forms in an attempt to be defensive about weird stuff in
wordlists.

We do still bake in the assumption that each grid cell can be
represented as one `char` -- this seems reasonable, but might need to be
revisited later.