Fix Unicode bugs in CLI wrapper
These are the same kinds of bugs that 18bedecfb fixed in the actual
solving code, but I forgot to look at the CLI wrapper. This fixes an
issue where the input grid would be rejected for the lines supposedly
not being the same length if it contained any multibyte characters.