Don't strip non-whitespace characters
It seemed like a good idea to normalize out punctuation when loading
word lists, but in practice this doesn't seem to be a big issue, and it
creates problems when the user does actually want to put punctuation in
the grid. Stripping whitespace is probably good enough.

The one qualm I have here is that if a wordlist wanted to include
apostrophes in the canonical versions of entries like "fool's gold"
(which I'm in favor of as a general principle), this would include the
apostrophe in the normalized version, making the word unusable. I might
want to make this behavior customizable in the future to deal with cases
like this.